Complete rewrite of the lexer pipeline.
Lexical analysis now takes place in a single pipeline, rather than three
distinct phases. To accomplish this, a few new classes have been
introduced.

First, a SourceManager class has been introduced, which is responsible
for caching the contents of files and tracking the locations of tokens.
The SourceManager allocates 32-bit identifiers representing a range of
positions for each a "source unit" (a file or macro).

A SourceLocation is no longer a (file, line, column) tuple. Instead it
is now an offset from a SourceManager ID. This makes the AST much
smaller in memory, albeit makes computing the full tuple more expensive.
Token buffers store some decoded information to keep parsing fast - but
the AST always uses the compactly encoded SourceLocation.

The benefit of these changes is that we can trace a token from where it
was created all the way to where it was inserted, all the way to each
file that caused it to be included.

The Preprocessor now replaces Scanner, and provides the same interface.
It has a stack of Lexers, which are something in between the old
TextProcessor and Scanner. Lexers are responsible for interpreting
tokens out of a source stream. They handle preprocessor directives, and
call back into the Preprocessor to handle difficult cases. For example,
the "#include" directive requires pushing a new Lexer, and sometimes
identifiers can require macro expansion.

Macros are now expanded in the same way Clang handles them. Rather than
lexing preprocessed text, we pre-lex #defines into a vector of tokens.
When we need to expand a macro, we push a much simpler MacroLexer, which
simply dishes out tokens from this vector.

Finally, we now track the extents of comments, though we don't do
anything with them yet.