Fix over-zealous cyclicy checks in typedefs.
It is not enough, unfortunately, to rely on type-resolution-recursion
for typedefs. We fail to parse something valid like:

   typedef A = function (A a);

Instead, typedefs now act like other AST nodes in that they compute an
empty initial type, and the TypedefType placeholder state has been
revived. We inline recursion checks as we do manually elsewhere and beef
up sizeof() to detect unresolved typedefs.

This lets us formally distinguish between types that are recursive due
to infinite name resolution, and types that are recursive due to
infinite size expansion.