Make name binding inline to parsing.
This is intended to simplify and speed up steps before semantic analysis
by eagerly computing as many components of the AST as we can. This both
lets us avoid storing extra state in the AST and lets us avoid full
walks of the AST in later passes.

The Parser now has a delegation API for hooking into important events,
and the NameBinder uses this to build the symbol table and bind names
inline.