Remove the heap tracker.
When we initially implemented the "heap tracker", the HEA section was relatively unused, except for
temporary native and argument storage space. The heap tracker was a *third* LIFO stack in the
system, that tracked the sizes of dynamically sized allocations into HEA. We use it for
dynamically-sized arrays and for an edge case in the ternary operator.

Why did we make a new LIFO stack instead of inlining the sizes into HEA? At the time, I figured HEA
was too easily corrupted, and it would be safer to track the allocations separately. In retrospect
this does not make much sense. The third structure is very difficult to inline in the JIT, it was
never properly validated or capped, and the HEA of course should only ever be internally
corrupted so the argument never made sense.

This patch removes the HeapTracker data structure, which predates SourceMod 1.0, and simply stores
allocation sizes directly into HEA.