Simplify the requirements of RETN.
SourcePawn has always required a balanced STK and HEA before a RETN instruction. This is
unnecessary, since the VM already knows the correct values, and indeed validates them anyway. If at
any point an exception is thrown, validation is skipped and the known-good values are restored.

It is much simpler to make RETN responsible for restoring the correct values. The compiler then
doesn't have to generate STACK and HEAP instructions before every return, or be burdened with
flowing all return paths through a common cleanup epilogue.

This adds 1-2 extra instructions on the generated prologue and epilogue paths, however, a compiler
that takes advantage of the new semantics can omit the final STACK/HEA balancing sequence.

For compilers to rely on the new semantics, they need a compatible VM, so the pcode version is
being bumped from 11 to 12.