Unify and simplify the test harness.
The test harness was too complicated. We had two versions, one for spcomp1 tests, and one for the
shell. In addition they were invoked by a weird script generated by the build, based on templates,
with separate templates for Windows and Mac/Linux. There's no need for any of this.

The new harness is much simpler. It scans an objdir for spcomp/spshell binaries (since we could
have dual-arch builds), scans for tests, and then runs a complete testing matrix. It now supports
per-folder manifests as well.

This setup will make it easier to add new configurations and even run tests in parallel.