Implement RTTI for functions, enums, typedefs, and typesets.
This is a major overhaul of the SMX layout. The following new tables are introduced:

 - `rtti.data`
 - `rtti.enums`
 - `rtti.methods`
 - `rtti.natives`
 - `rtti.typedefs`
 - `rtti.typesets`

These tables encode information that will prove to be essential in making
Pawn type-safe. They are also the gateway to many language improvements. As
such, these tables are considered to be mandatory even if their use is minimal
right now. The VM will depend on them in the future.