SmxViewer: Collect unknown functions while disassembling
When encountering a `call` while disassembling check if there is an entry for the called address in the ".publics" or ".dbg.symbols" sections. If the function is missing add it to a seperate list to be able to disassemble functions even while debug info is stripped.

The functions are added with a name like `sub_<addr>` to be unqiue.