Remove cmptag.
This doesn't exist in upstream Pawn. It's awkwardly attached to value
structs even though it's actually parsing state. Even more strange is
that sometimes it's a tag and sometimes it turns into a symbol count.

It allowed an awkward feature to exist where enum members could be
duplicated or shadowed across enum declarations. The compiler would then
attempt to deduce the correct declaration when used as an index. Having
shadowed constants in the global namespace is bizarre and we're better
off not having this feature, or even trying to replicate it through
saner parsing tricks.

Now, if enum fields are shadowed, the compiler will error. The
short-term fix for this is for developers to prefix their enum fields
(which is common practice in core code).