Rewrite the lexer.
lex() is pretty unreadable, it has duplicated code and weird state
transitions, and it's too long. This patch breaks it up into separate
methods that each perform a clear task.

The operator lookup code is now inlined into a switch statement (similar
to how the v2 lexer works), removing a loop+strcmp over a sizable chunk
of the token table. Compiling SourceMod's plugins looks about 5% faster
with this change.

After this change, it will be much easier to implement a token map and
remove the sc_token loop.