Refactor symbol and keyword lexing.
Previously, the lexer would consume characters until it matched a
keyword, and would then commit the new stream pointer. Instead, we
should consume characters up until an entire lexeme is available, and
then deduce whether it is a keyword or symbol.

This unfortunately is a bit tricky since there are lots of special
cases:

  '@' is equivalent to 'public' if alone.
  '_' is a special token if alone.
  '#' can begin a keyword but not a symbol.
  '@' can begin a symbol but not a keyword.
  Symbols can contain numbers, but keywords cannot.

The new logic handles these cases as the old code did.