Remove memfile_gets.
The output assembly listing must be parsed twice for code generation.
This is currently done through an awkward gets() implementation on the
in-memory stream, with some clumsy string manipulation code afterward.

This patch removes memfile_gets(), and forces the code generator to use
a mini-lexer to parse the stream instead.