Add transitional syntax for enum structs.
This amends the Transitional Syntax to allow enum structs, codifying a
permanent replacement and obsoleting the old syntax. The grammar is:

    enum struct <name> { <newline>
      [<new-type-expr> <name> [post-dims] <newline>]*
    } <newline>

Enum structs declared this way can be used anywhere a new-style type is
declared. Internally, they still decompose into arrays. As such,
anywhere an array is forbidden, enum structs are not allowed.
Similarly, if used in an array, an enum struct type will become a
multi-dimensional array.

Unlike old-style enum structs, they are not accessed with array syntax.
Instead they are accessed through the field operator ("."). For example
an old enum struct might look like:

    enum Sample {
      Float:x, Float:y
    };
    new s[Sample];
    s[x] = 10; s[y] = 20;

A new-style example might look like:

    enum struct Sample {
      float x;
      float y;
    };
    Sample s;
    s.x = 10; s.y = 20;

Similarly, the field names of a new-style enum struct do not poison the
global scope. In fact they are not accessible, and the enum struct name
itself does not define a global symbol (only a named type).

Additionally, a long-standing bug has been fixed where strings within
enum structs could not be indexed properly. Also note that enum structs
are forbidden in natives, and must be casted to any[] to be communicated
across plugin boundaries. If doing this cross-plugin, care must be taken
that the ordering of the enum struct fields do not change.

This new syntax obsoletes old-style enum structs. They will no longer be
maintained going forward, and bugs in them will not be fixed. They may
be broken in the future.

Rationale:

Enum structs have been neglected due to having a painfully complex and
strange implementation. Its problems roughly fall into three categories:
 1. Syntax: they are awkwardly layered on top of enums and arrays.
 2. Semantics: they define unscoped globals that expose implementation
    details, and can be used in illegal, nonsensical ways.
 3. Implementation: rather than define a type, they create a shadow
    structure attached to arrays, which then requires complex pattern
    matching to reconstruct.

Developers have been asking for struct syntax, even if it's sugar around
enum structs. In the past we've been reluctant to do this, because
Structs remain a large carrot for moving to a newer compiler. Properly
implemented, they allow many things enum structs do not (including
cross-plugin and native support). Syntactic sugar does not fix the deeper
problems in the runtime, and to fix those, we would need users to accept
backwards compatibility breaks and experimental behavior. Thus, structs
would be the carrot to entice this.

However this thinking was probably mistaken. No "correct" struct
implementation has been forthcoming, and meanwhile, we're stuck with a
poorly implemented and widely used language misfeature. It is better to
provide something maintainable to promote better plugins, and provide
developers with a better environment. This solution is also incomplete:
there are many things enum structs still don't allow, so we have not
lost a carrot for future versions.