Disallow reference tokens on enum struct parameters.
This essentially guarantees that future struct implementations will have
pass-by-reference semantics. The alternative would be requiring a
reference token for enum structs, and then leaving the door open for
pass-by-value semantics later. However, there do not seem to be many use
cases for pass-by-value that could not be solved in other ways. And this
implementation matches how arrays and methodmaps are already used.