Fix issues related to line terminators on global declarations.
This patch tweaks the require_newline algorithm to allow for three
scenarios:

(1) A declaration where ONLY newlines are allowed to terminate a
statement. Enum structs now fall into this category.

(2) A declaration where semicolons OR newlines are allowed to terminate
a statement. Methodmaps, typesets fall into category. If a semicolon is
missing, requiring semicolons with a pragma will not error. This mode is
for compatibility, since the original semantics were a mistake
(methodmaps should be "newline", typesets should be "semicolon").

(3) A declaration where semicolons are required to terminate a
statement. If semicolons are not required with a #pragma, then the
missing semicolon is ignored and a newline is accepted.

Finally, this patch fixes a bug where token_pos_t did not track the file
number of the token, leading to incorrect error positions if the lexer
jumped past an #include directive.

Bug: issue #314.