Ensure that methodmaps always have a name.
When we encounter an unknown method call, we will try to generate a
temporary symbol with a made-up name. If the methodmap has no name,
this will result in a symbol starting with ".", for example,
".SetValue". When finding the display name for this function, the
compiler assumes that a name starting with a non-identifier symbol
is special, and must therefore be an operator. Since operator names have
a different maximum length than normal symbol names, this causes a stack
overflow.

In addition this patch fixes the function that can erroneously overflow.

Bug: fixes issue #333
Test: compile-only/fail-methodmap-keyword-name