Simplify inline exit frames.
When invoking C++ from JIT code, we need to construct an "exit frame"
so that FrameIterator can recover a stack trace. The normal way of
creating an exit frame is implicitly, via a call. The call pushes a
return address, and the new prologue configures the frame pointer and
stores it in the Environment.

Notably, this means there is a small inline helper function every time
we transition to C++. For performance-sensitive code, like native calls,
this is not ideal. Instead we create an "inline" exit frame. Instead of
calling, we derive the return address at compile-time, push it, and thus
avoid the call instructions.

This patch makes additional improves to the inline-exit path by also
eliding "leave" instructions. It also preserves ebp, which should
improve stack walking by Breakpad.