Create actual symbols for enum struct fields. (#353)
hier1() had an ancient hack that created a dummy symbol for enum struct
fields. This dummy symbol was stack allocated, so we couldn't propagate
it through the expression parser. This attaches the symbols to the
variable so they can escape, which gives the compiler much richer
information about the field. Note that these one-off symbols could be
re-used, but we do not re-use them in this patch.

Bug: issue #301
Test: tests/enum-structs