Eliminate the native counter hack.
For some reason, natives had a weird relationship with the assembler.
The code generator assigned an address based on a global counter, and
then the assembler was responsible for building the table in the same
order. (Similar nonsense exists with the staging counter, but we'll save
that for another day.)

Instead, emit sysreq instructions by name as we do for the call
instruction. This allows us to build the table during code generation,
simplifying another piece of the assembler pipeline.

Bug: N/A
Test: manual test