Remove legacy enum structs.
This patch effectively cuts enum structs off at the knees. The root constant
created by an enum is now defined as 0, and it's illegal to create
arrays using this constant. Parsing support has been removed from the
index operator, and the index operator (and old-array decl syntax) now
require types implicitly convertible to integers.

In addition, support for tagging items in enums has been removed, as
there is no longer a need for this behavior.

The underlying mechanism for this mess is still supported, however,
because it is used for transitional enum structs. This patch simply
removes the worst hacks in the parser and hier1().

Plugins using the old array-as-struct syntax will no longer compile, and
should use the new enum struct syntax instead.

Bug: N/A
Test: new tests