Fix bugs in the increment/decrement operator.
The post-incdec operator is particularly broken in the old code
generator. We make no attempt to fix it here, since it is not worth
trying to preserve old behavior in this case. The new parser avoids
calling emit_userop since it is complicated and inefficient.

New tests for this will fail on the current parser, so they are
disabled.

Unfortunately, the new code generation patterns wildly trip up the
optimizer. The optimizer has many patterns that result in incorrect
replacements (for example, exchanging a move+push for a push). Since
the long term plan is to remove the optimizer and simply emit better
code, we now disable the optimizer for expressions using the new parser.