Fix marking function return values.
The old parser could not determine whether a function call result was
used or not. To implement the warning, it would aggressively check
whether the inputs to an expression involved a function symbol, and mark
it as needed a return value. This mostly worked because check was copied
into the most common cases in the parser.

This broke in the new parser because functions identifiers are pipelined
rather than immediately evaluated. They always return a function symbol,
and the caller is responsible for evaluating it. This is different from
the old parser, where a function symbol immediately created a call
expression or a "function value" expression. As a result, the hack for
aggressively poking at potential function calls no longer works.

This can be resolved by using the AST. Each node must implement a
ProcessUses() call. For each child expression whose value is consumed,
it must call MarkUsed(). For CallExpr, this will mark the function as
needing a return value.

Bug: issue #402
Test: compile-only/ok-void-function-as-value.sp