Fix constants being handled incorrectly with commutative operator overloads.
The old expression parser would actually swap value structs here, so
find_userop would implicitly swap parameters. This detail did not carry
into the new parser, because the semantic and codegen steps are now
separate. The semantic analysis step determines the overload signature,
and by the time we reach code generation, it is too late to pretend the
semantic values are swapped (this would require choosing a new
overload).

A quick fix is to force swapparams=true on the UserOperation when we make
this optimization. Longer term, this logic will be replaced. The
experimental compiler has a bunch of improvements to how PRI/ALT are used
that we want to backport to the production compiler.

Bug: issue #424
Test: tests/basic/constval-user-op