Process string literals in a staging buffer.
Currently, strings are processed in an idiosyncratic way. First, each
segment of the string is scanned into the token's buffer. Then, the
entire buffer is lexed character by character, and as this happens,
characters are added to the literal queue. This means it is impossible
to lex a string without dumping bytes directly into the data blob, and
in addition, raw and non-raw strings cannot be mixed, because the flags
are "global" for the string rather than each segment.

This patch begins to undo the mess by scanning strings into a temporary
buffer, and then lexing the characters into the token. The final string
is then added to the literal queue.