Remove support for RAWMODE.
RAWMODE was the ability to prefix a string such that escape characters
would be inserted literally, similar to the "r" prefix in Python. It
used sc_ctrlchar as the prefix. For reasons unknown, I requested this
feature be disabled when literal concatenation support was backported [1].
Since then, it was further broken when the lexer was rewritten and
simplified. It hasn't worked in 6 years, and no one has been clamoring
for it, so let's just remove all the offending code.

If we want to add this back and properly support it, it will not be
difficult.

[1] https://github.com/alliedmodders/sourcemod/pull/5