Force callers to litadd() string literals.
This removes a long-standing egregious boundary violation in the lexer:
it would implicitly add to the data segment if it encountered a string.
This had all sorts of bizarre implications, and also necessitated a hack
(AutoDisableLiteralQueue) to ensure deep peekahead didn't break other
parts of the compiler.

Forcing callers to manually litadd() a string eliminates all of these
quirks, and we can now remove the AutoDisableLiteralQueue hack. Doing so
was made possible by the preceding refactorings.

This will also enable further splitting of code generation and parsing.