Move p-struct declarations to the new parser.
In addition to moving declstructvar() into the new AST-based parser,
this patch has some semantic and implementation changes worth noting.

P-struct semantics have been narrowed to precisely match their use case.
They will be expanded when the syntax for enum structs and non-enum
structs can be unified. As part of this, non-string arrays, references,
and variables are no longer allowed anywhere in a p-struct definition.
Type checking is performed on field assignment now as well.

In addition, dumplits() now resets the literal queue. The literal queue
is (like most things in spcomp) a hack. It allows resetting the state of
DATA while skipping functions, and rewriting indirection vectors of 3D
arrays. When these use cases are removed, the literal queue will
eventually be removed as well. For now, we make it easier to use by
ensuring the queue is empty after being dumped.

Finally, we no longer mark public vars as "read". Instead, the public
flag forces inclusion into the pubvars and .dbg.globals tables, to match
the behavior of public functions.