Include tag names in tag mismatch error message
I know this isn't quite perfect as the type resolution can be quite
complex, and it is generally better to provide less information than
bad information, but I think this improves DX in the vast majority
of cases and is probably worthwhile doing unless there is anything
super obviously wrong.

A lot of users have trouble understanding this error message, so a
little bit of added context like this should go a long way. I've
kept the message as "tag mismatch" instead of "type mismatch" for
now, but that might be worth changing at some point as well. I've
also changed "loose intendation" to clarify what it is checking a
little better, as that is another common warning that people don't
understand.