Change how dynamic natives are created.
Dynamic natives are an important SourceMod feature, but they are very
burdensome to implement in the VM. They're a holdover from AMX Mod X and
require a platform-specific assembly blob. Every time we add support for
a new architecture, we have to figure out how to implement this on every
possible ABI. On platforms where we're never going to add a JIT, it
means implementing parts of the JIT assembly framework for no reason.

This patch removes CreateFakeNative. Embedders must now call
UpdateNativeBindingObject() which takes in a reference counted,
user-specified object, allowing the embedder to customize behavior.

Embedders (SourceMod included) using the old API will no longer
function.