Improve macro recursion support.
The previous fix for this was intended to be a depth test, but actually
limited the number of macros per line, since substitution is iterative
and not recursive.

Instead, we now allow substituting macros infinitely. However each
substitution begins a "sequence", and this sequence is extended as new
substitutions are made. Any macro used inside this sequence is banned
from being substituted again. Once there are no more substitutions in
the sequence, a new one can begin, and the banned set is cleared.

This means that "#define TEST TEST" will now insert the characters
in "TEST", rather than infinitely perform substitutions, or return a
fatal error. This matches the C preprocessor.

Bug: issue #526
Test: ok-recursive-macro.sp