Don't allow compiler symbols in preprocessor expressions.
This was always a super weird quirk, so let's wean the language off it.
The preprocessor is textual and is a completely different language
versus actual SourcePawn semantics. It never made any sense that it
could reach into language constructs.

Most likely this was accidental and not intentional. It was easier for
CompuPhase to re-use the "const" parser (used for switch cases and
stuff), which pulls in enums and const variables as well.

Since the damage here is unknown, we use a warning instead of an error.
It will become an error in the 1.12 release.

Bug: issue #561
Test: CI