Add a new static_assert statement.
The preprocessor directive `#assert` can only reference preprocessor
constructs. There is no way to assert compile-time constants defined in
SourcePawn.

This change introduces the `static_assert` statement, which has two
forms:

	static_assert(expression)
	static_assert(expression, "message")

The expression must be constant, eg, function calls are not allowed, and
any identifiers must be enum identifiers or constant variables. The
message is optional, but if provided, it must be a string literal.
Variables are not allowed.

If the expression evaluates to zero or false, an error is raised and
compilation will fail.

`static_assert` is only valid at global scope.

Grammar:
	static-assert ::= "static_assert" "(" const-expr ("," string-literal)? ")"

Bug: #561
Test: new compile-only tests