Fix inconsistencies with function signature type checks.
This fixes a few bugs in typedef/function signature checks.

First, typedef-to-typedef comparisons are now valid. They previously
would always fail, due to the type checker requiring the type to be
backed by a function symbol.

Second, methodmap checks are now bidirectional. Eg, Handle casts to
DataPack and DataPack casts to Handle. This is a convenience cludge that
is commonly used, we can't easily break it, so now it is codified.

Third, arrays only implicitly cast if tags are identical. This is not
ideal, but it suffices until better rules can be written. For now it
prevents char[] casting to something with integer width.

Finally, a callback using a non-"any" type cannot be passed to a callback
using the "any" type. This prevents illegal implicit casts that we can't
yet do runtime checks on to prevent.

Bug: #548
Test: new compile-only tests