Require RTTI-enabled plugins to have valid RTTI entries.
This validates that all method calls have an associated RTTI entry, if
the RTTI table is present. Signature validation will come in 1.12.

Opcodes do not yet depend on RTTI, so the compiler does not yet enforce
RTTI tables to be emitted. This is to allow 1.11-compiled plugins to run
on SourceMod 1.10.

Bug: issue #580
Bug: alliedmodders/sourcemod #1525