Fix enum names not being visible in enum declarations.
The new parser abuses exprconst() to avoid dealing with constant
evaluation of expressions. We don't need to do this though, as Analyze
already does most of our constant folding. To walk back our dependence
on calc() however we do re-implement constant folding for BinaryExpr,
and constant folding should be moved further into FoldToConstant() as we
deprecate the old parser.

Bug: #509
Test: new runtime test