Add a tool for testing a plugin corpus.
When making major changes to the compiler, we usually test them against
a saved copy of all plugins on the forums. Add a script to help automate
this.

Use `--remove-good` to reduce a corpus to only breaking plugins. This
helps narrow down a worklist.

Use `--remove-bad` to reduce a corpus to only working plugins. This
helps weed out known-bad plugins on a good compiler.

Use `--diagnose` to interactively choose whether to remove or keep a
non-working plugin.

Note that about half of all plugins in the current corpus will fail even
on a good compiler. We do not download includes or zip files so we'll
often miss dependencies. Still, this has been a pretty effective method
and gives us a test base of around 9,000 source files.