Move local statement handling to the AST pipeline.
This guts parser.cpp and moves all local statement parsing to the new
parser. The change is pretty straightforward, though it's a lot of code
moving around.

Control-flow analysis can now be a lot more flexible. It's easier to
detect conditions like missing returns, endless loops, improve
code-generation, and constant-fold if/loop conditionals.

The biggest internal change is that the compound/nesting level has been
removed in favor of a traditional scope chain hierarchy. Eventually
loctab and glbtab will be removed entirely. This change has greatly
simplified code which works with the local symbol table.

Error reporting has been tweaked slightly. The old "one error per
statement" concept has been retained for parsing, but is not used for
binding or semantic analysis. This may change in the future, but very
erroneous (but syntactically correct) lines may be more spammy now.

Another change is that it is now an error to mix "return" and "return
<value>" statements, or to omit return statements in new-style
functions. A few exceptions have been made to limit imapct, and in these
cases, the error becomes a warning instead.

Symbol handling is slowly moving into symbols.cpp, and parser.cpp will
be renamed to main.cpp once the old parser is gone.

Reliance on global variables is going to dwindle from here on as well.
New analysis bits will be rooted on SemaContext (and Parser). Long-term
this will allow us to compile on-demand within srcds in parallel
threads.